Fix #242 Number input does not show validation errors if you enter text #242
The major browsers do not pass on a invalid number input as value to the javascript. Instead a empty value is passed. This blocks our validators from validating the invalid chars and creates a difference between what the browser displays and what stored as field value ( 'one' is stored as null).

As fix/work-around we can check the validity of the input ( supported by most, but not all browsers )  on keyup and if the inout is not valid put back the value as it was set before the input event.