fix: Show error message for number inputs if html validation hides the actual value
Html validation puts the value to null even if the user sees content.
This is an issue because then the vue-form validation cannot do its thing.

To detect this case, add a custom message if the value is empty.
Then check for two pseudo classes in css to determine if the message should be visible.
* If the html validation fails the input has the :invalid pseudoclass.
* If the field is empty, the input has the :placeholder-shown pseudoclass.
We need to set an empty string placeholder for that to work.

While we're at it, also show the validation messages in red.

Fixes #242, Fixes #245