fix: add N/A option when radio field is nillable
The EntityToFormMapper currently adds an N/A value for enums and booleans with attribute.nillable.
Instead, let RadioFieldComponent add the N/A value if the attribute is currently required.
As a bonus, this also takes into account the attribute.nillableExpression.
For backward compatibility, use form_boolean_missing for all attribute types.

Fixes #184 (for the categoricals)